chore(Service): Make code more readable

+++ src/main/java/com/bank/vote/voteitem/VoteItemService.java
@@ -1,9 +1,10 @@
 package com.bank.vote.voteitem;
 
+import com.bank.vote.common.Exceptions.ValueNotAllowException;
 import com.bank.vote.common.Exceptions.VoteItemAlreadyExistedException;
+import com.bank.vote.common.Exceptions.VoteItemNotFoundException;
 import com.bank.vote.voterecord.VoteRecordRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Propagation;
@@ -30,6 +31,9 @@ public class VoteItemService {
 
     @Transactional(readOnly= false, isolation = Isolation.READ_COMMITTED, propagation = Propagation.REQUIRED)
     public VoteItem addVoteItem(VoteItem voteItem) {
+        if (!isValidItemName(voteItem.getItemName())){
+            throw new ValueNotAllowException("ItemName not Allow!!");
+        }
         Optional<VoteItem> existingVoteItem = voteItemRepository.findVoteItemByItemName(voteItem.getItemName());
         if (existingVoteItem.isPresent())
             throw new VoteItemAlreadyExistedException("Item already existed.");
@@ -37,8 +41,16 @@ public class VoteItemService {
     }
 
     @Transactional(readOnly= false, isolation = Isolation.READ_COMMITTED, propagation = Propagation.REQUIRED)
-    public void removeVoteItem(Integer itemId) throws ChangeSetPersister.NotFoundException {
+    public void removeVoteItem(Integer itemId) {
+        if(voteItemRepository.findById(itemId).isEmpty()){
+            throw new VoteItemNotFoundException("Item not Found.");
+        }
         voteRecordRepository.deleteByVoteItemId(itemId);
         voteItemRepository.deleteById(itemId);
+    }
+
+    private boolean isValidItemName(String itemName) {
+        String regex = "^[\\p{L}A-Za-z0-9]+$";
+        return itemName.matches(regex);
     }
 }
